pyproject.toml
[tool.poetry]
name = "zametki-tg"
version = "0.1.0"
description = ""
authors = ["diglander <denisovchinn2094@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
pydantic = "^2.12.5"
sqlalchemy = "^2.0.45"
syncpg = "^1.1.3"
alembic = "^1.18.1"
celery = "^5.6.2"
redis = "^7.1.0"
python-dotenv = "^1.2.1"
groq = "^1.0.0"
openai = "^2.15.0"
psycopg2-binary = "^2.9.11"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
ruff = "^0.14.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
# E/W - стиль, F - ошибки, I - импорты, B - плохие практики
select = ["E", "W", "F", "I", "B"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
